backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
channels==3.0.5
charset-normalizer==3.4.7
cloudinary==1.44.1
Django==3.1.12
django-cloudinary-storage==0.3.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djongo==1.3.7
dnspython==2.7.0
gunicorn==23.0.0
idna==3.11
packaging==26.0
pillow==11.3.0
PyJWT==2.12.1
pymongo==3.11.4
python-dotenv==1.2.1
pytz==2026.1.post1
requests==2.32.5
six==1.17.0
sqlparse==0.2.4
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.11.0
